requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
Flask>=3.0.0
waitress>=3.0.0
qrcode>=7.4
Pillow>=10.0.0
